feat(web): unify plan hub with nested calendar and packing routes
