+++ apps/web/app/plan/packing/[id]/page.tsx
@@ -9,7 +9,6 @@ import type { Doc, Id } from "@convex/_generated/dataModel";
 import { format, eachDayOfInterval } from "date-fns";
 import { useRequireAuth } from "@/hooks/use-require-auth";
 import { PageContainer } from "@/components/layout/page-container";
-import { UserAvatar } from "@/components/user-avatar";
 import { BrutalistButton } from "@/components/brutalist-button";
 import { useOutfitRecommendations } from "@/hooks/use-outfit-recommendations";
 import type {
@@ -65,10 +64,10 @@ function convexGarmentToApi(g: Doc<"garments"> & { _creationTime?: number }): {
   };
 }
 
-export default function PackingTripPage() {
+export default function PlanPackingTripPage() {
   const params = useParams();
   const id = params.id as Id<"packingLists">;
-  useRequireAuth("/packing");
+  useRequireAuth("/plan/packing");
 
   const trip = useQuery(api.packingLists.get, { id });
   const allGarments = useQuery(api.garments.list) ?? [];
@@ -271,106 +270,73 @@ export default function PackingTripPage() {
 
   if (trip === undefined) {
     return (
-      <main className="min-h-screen bg-background text-foreground">
-        <header className="fixed top-0 left-0 right-0 z-50 glass-bar rounded-none border-x-0 border-t-0 border-b border-border">
-          <div className="flex items-center justify-between px-4 py-5 md:px-8 lg:px-10 xl:px-12">
-            <Link
-              href="/packing"
-              className="text-[10px] md:text-xs uppercase tracking-[0.3em] font-medium hover:text-signal-orange transition-colors duration-100 cursor-pointer"
-            >
-              Packing
-            </Link>
-            <UserAvatar />
-          </div>
-        </header>
-        <div className="pt-24 pb-24">
-          <PageContainer>
-            <Skeleton className="h-6 w-56 mb-4" />
-            <Skeleton className="h-4 w-72 mb-10" />
-            <Skeleton className="h-40 w-full mb-8" />
-            <Skeleton className="h-32 w-full max-w-md" />
-          </PageContainer>
-        </div>
-      </main>
+      <PageContainer>
+        <Skeleton className="h-6 w-56 mb-4" />
+        <Skeleton className="h-4 w-72 mb-10" />
+        <Skeleton className="h-40 w-full mb-8" />
+        <Skeleton className="h-32 w-full max-w-md" />
+      </PageContainer>
     );
   }
 
   if (trip === null) {
     return (
-      <main className="min-h-screen bg-background">
-        <div className="pt-24 pb-24">
-          <PageContainer>
-            <p className="text-[11px] uppercase tracking-[0.2em] text-muted-foreground">
-              Trip not found.
-            </p>
-            <Link
-              href="/packing"
-              className="mt-4 inline-block text-signal-orange text-sm uppercase tracking-widest cursor-pointer hover:underline"
-            >
-              Back to packing
-            </Link>
-          </PageContainer>
-        </div>
-      </main>
+      <PageContainer>
+        <p className="text-[11px] uppercase tracking-[0.2em] text-muted-foreground">
+          Trip not found.
+        </p>
+        <Link
+          href="/plan/packing"
+          className="mt-4 inline-block text-signal-orange text-sm uppercase tracking-widest cursor-pointer hover:underline"
+        >
+          Back to packing
+        </Link>
+      </PageContainer>
     );
   }
 
   return (
-    <main className="min-h-screen bg-background text-foreground selection:bg-signal-orange selection:text-background">
-      <header className="fixed top-0 left-0 right-0 z-50 glass-bar rounded-none border-x-0 border-t-0 border-b border-border">
-        <div className="flex items-center justify-between px-4 py-5 md:px-8 lg:px-10 xl:px-12">
-          <Link
-            href="/packing"
-            className="text-[10px] md:text-xs uppercase tracking-[0.3em] font-medium hover:text-signal-orange transition-colors duration-100 cursor-pointer"
-          >
-            Packing
-          </Link>
-          <UserAvatar />
-        </div>
-      </header>
+    <>
+      <PageContainer>
+        <TripPageHeader
+          tripName={trip.name}
+          startDate={trip.startDate}
+          endDate={trip.endDate}
+        />
 
-      <div className="pt-20 sm:pt-24 md:pt-28 lg:pt-32 pb-24 md:pb-28">
-        <PageContainer>
-          <TripPageHeader
-            tripName={trip.name}
-            startDate={trip.startDate}
-            endDate={trip.endDate}
+        <section className="mb-8">
+          <div className="flex items-center justify-between gap-4 mb-3">
+            <h2 className="text-sm uppercase tracking-widest">
+              Packed pieces ({trip.garmentIds.length})
+            </h2>
+            <BrutalistButton
+              variant="outline"
+              size="sm"
+              onClick={() => setClosetModalOpen(true)}
+              className="cursor-pointer"
+            >
+              Add from closet
+            </BrutalistButton>
+          </div>
+          <PackedGarmentsGrid
+            garments={packedGarments as Doc<"garments">[]}
+            onRemove={removeFromTrip}
           />
+        </section>
 
-          <section className="mb-8">
-            <div className="flex items-center justify-between gap-4 mb-3">
-              <h2 className="text-sm uppercase tracking-widest">
-                Packed pieces ({trip.garmentIds.length})
-              </h2>
-              <BrutalistButton
-                variant="outline"
-                size="sm"
-                onClick={() => setClosetModalOpen(true)}
-                className="cursor-pointer"
-              >
-                Add from closet
-              </BrutalistButton>
-            </div>
-            <PackedGarmentsGrid
-              garments={packedGarments as Doc<"garments">[]}
-              onRemove={removeFromTrip}
-            />
-          </section>
-
-          <OutfitIdeasTripSection
-            mood={mood}
-            onMoodChange={setMood}
-            onGenerate={handleGenerate}
-            loading={loading}
-            error={error}
-            displayOutfits={displayOutfits}
-            tripDates={tripDates}
-            onSaveLook={handleSaveLook}
-            onAssignToDay={handleAssignToDay}
-            generateDisabled={packedForApi.length === 0}
-          />
-        </PageContainer>
-      </div>
+        <OutfitIdeasTripSection
+          mood={mood}
+          onMoodChange={setMood}
+          onGenerate={handleGenerate}
+          loading={loading}
+          error={error}
+          displayOutfits={displayOutfits}
+          tripDates={tripDates}
+          onSaveLook={handleSaveLook}
+          onAssignToDay={handleAssignToDay}
+          generateDisabled={packedForApi.length === 0}
+        />
+      </PageContainer>
 
       <ResponsivePlanModal
         open={closetModalOpen}
@@ -399,6 +365,6 @@ export default function PackingTripPage() {
           onCancel={() => setClosetModalOpen(false)}
         />
       </ResponsivePlanModal>
-    </main>
+    </>
   );
 }
